KDllAr: convert AOUT to OMF before emxexp
This fixes the problem that a long symbol name is not found in .def
because emxomf fits a long symbol into 255 chars which is a limitation
of OMF.

That is, emxexp lists the symbol name not converted by emxomf. But when
being linked, a long symbol name modified by emxomf is used. As a
result, a long symbol name listed by emxexp cannot be found.

Fixes #2.

    modified:   kdllar.cpp
    modified:   kdllar.h