KDllAr: check if eof as well as if good
If a .o file has only a header, ifstream::read() cannot read up to
`sizeof( hdr )' which is `sizeof( ar_hdr )'. Because of this,
ifstream::read() clears a good bit and set a eof bit.

Therefore, it's necessary to check a eof bit as well as a good bit.

This problem occurs when converting common/x86/pixel-32-10.o to build
x264164.dll of libx264.

    modified:   kdllar.cpp